bpm-analyzer/requirements.txt
# Core dependencies
numpy>=1.24.0
madmom>=0.16.1
librosa>=0.10.0
jams>=0.3.4
soundfile>=0.12.0
pydub>=0.25.1
typer[all]>=0.9.0
rich>=13.0.0
pydantic>=2.0.0
sqlalchemy>=2.0.0
alembic>=1.12.0
pyyaml>=6.0

# Optional dependencies (uncomment as needed)
# essentia>=2.1b6
# aubio>=0.4.9
# psycopg2-binary>=2.9.0
